Move specs into spec/lib to reflect top-level structure
