+++ spec/lib/index_spec.js
@@ -1,4 +1,4 @@
-const router = require('../lib/index');
+const router = require('../../lib/index');
 const expect = require('chai').expect;
 const sinon = require('sinon');
