+++ spec/lib/pairrit_spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const pairrit = require('../lib/pairrit');
+const pairrit = require('../../lib/pairrit');
 const SHA256 = require('crypto-js/sha256');
 
 describe('Pairrit', function() {
